+++ src/containers/PrivatePage/PrivatePage.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import Image from '../../components/Image';
-import FooterSpot from '../../modules/FooterSpot';
+import Footer from '../../modules/Footer';
 import LoginHeader from '../../modules/LoginHeader';
 import data from './data.json';
 
@@ -20,16 +20,16 @@ export default function PrivatePage(): JSX.Element {
         <div className="blur-theme">
           <div className="btn-container">
             <div className="btn-position">
-              <button className='red-btn' type='button'>{BTN_CONTENT}</button>
+              <button className="red-btn" type="button">
+                {BTN_CONTENT}
+              </button>
             </div>
             <div className="btn-line">.</div>
           </div>
           <div className="movie-precontainer">
             <div className="movie-container">
               <h1 className="movie-title">{data.movieTitle}</h1>
-              <h3 className="movie-meta-info">
-                {`${data.item} ${data.releas} ${data.country} ${data.date}`}
-              </h3>
+              <h3 className="movie-meta-info">{`${data.item} ${data.releas} ${data.country} ${data.date}`}</h3>
               <div className="movie-description">
                 <h6 className="movie-description__part-title">{data.loglineTitle}</h6>
                 <div className="movie-description__part">
@@ -40,7 +40,7 @@ export default function PrivatePage(): JSX.Element {
                   <p>{data.synopsis}</p>
                 </div>
                 <div className="movie-tags">
-                  {data.tags.map((item  ) => (
+                  {data.tags.map((item) => (
                     <div className="movie-tags__tag" key={`tag-${item}`}>
                       <span className="movie-tags__tag-wrapper">{item}</span>
                     </div>
@@ -48,19 +48,15 @@ export default function PrivatePage(): JSX.Element {
                 </div>
               </div>
             </div>
-            <div className="pre-movie-poster">  
+            <div className="pre-movie-poster">
               <div className="movie-poster">
-                <Image
-                  src={data.imgSrc}
-                  alt={data.alt}
-                  className='movie-img'
-                />
+                <Image src={data.imgSrc} alt={data.alt} className="movie-img" />
               </div>
             </div>
           </div>
         </div>
       </section>
-      <FooterSpot footerPrefix="login-page__footer" />
+      <Footer footerPrefix="internal-footer" />
     </>
   );
 }
