+++ src/modules/Footer/footer.scss
@@ -82,3 +82,42 @@ footer {
 .dev-link:hover {
   text-decoration: underline;
 }
+
+.internal-footer {
+  background-color: rgb(22, 22, 22);
+  padding: 18px;
+  display: flex;
+  flex-direction: row;
+  justify-content: flex-start;
+}
+
+.internal-footer__links {
+  width: 33%;
+  min-width: 300px;
+  margin-bottom: 36px;
+  margin-top: 10px;
+  display: flex;
+  flex-direction: column;
+  padding-left: 40px;
+}
+
+.internal-footer__item_descr {
+  color: rgba(255, 255, 255, 0.7);
+  font-size: 10px;
+  margin-bottom: 10px;
+}
+
+.internal-footer__item_link {
+  color: rgba(255, 255, 255, 0.9);
+  display: block;
+  padding: 10px 0;
+  font-size: 14px;
+  text-decoration: none;
+}
+
+.footer-icons__twitter,
+.footer-icons__instagram,
+.footer-icons__facebook {
+  fill: rgba(255, 255, 255, 0.9);
+  width: 24px;
+}
