+++ src/modules/Footer/index.tsx
@@ -1,0 +1,3 @@
+import FooterSpot from './Footer';
+
+export default FooterSpot;
