+++ src/modules/Footer/svg/InstargamSVG.tsx
