+++ src/modules/Footer/types.ts
