--- src/modules/FooterSpot/index.tsx
@@ -1,3 +1,0 @@
-import FooterSpot from './FooterSpot';
-
-export default FooterSpot;
