--- src/modules/InternalFooter/index.ts
@@ -1,3 +1,0 @@
-import InternalFooter from './InternalFooter';
-
-export default InternalFooter;
